+++ src/app/router_registry.py
@@ -2,6 +2,9 @@ from fastapi import FastAPI
 
 from app.routers.advisor_cockpit import router as advisor_cockpit_router
 from app.routers.advisory_policy import router as advisory_policy_router
+from app.routers.advisory_policy_evaluations import (
+    router as advisory_policy_evaluations_router,
+)
 from app.routers.advisory_workspaces import router as advisory_workspaces_router
 from app.routers.analytics_diagnostics import router as analytics_diagnostics_router
 from app.routers.archive_documents import router as archive_documents_router
@@ -127,6 +130,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(bank_demo_proof_router)
     app.include_router(advisory_workspaces_router)
     app.include_router(advisory_policy_router)
+    app.include_router(advisory_policy_evaluations_router)
     app.include_router(proposals_router)
     app.include_router(proposal_operations_router)
     app.include_router(proposal_versions_router)
